organized localstorage functions
